+++ all-is-cubes-gpu/tests/shaders/harness.rs
@@ -1,6 +1,3 @@
-use std::io::Write as _;
-use std::sync::Arc;
-
 use half::f16;
 
 use all_is_cubes::camera;
@@ -12,33 +9,16 @@ use all_is_cubes_gpu::in_wgpu::shader_testing;
 ///
 /// We don't share the [`wgpu::Device`] because it can enter failure states,
 /// but we can use just one [`wgpu::Adapter`] to create all of them.
-pub(crate) async fn adapter() -> Arc<wgpu::Adapter> {
-    static CELL: tokio::sync::OnceCell<Arc<wgpu::Adapter>> = tokio::sync::OnceCell::const_new();
-
-    CELL.get_or_init(|| async {
-        let (_instance, adapter) =
-            init::create_instance_and_adapter_for_test(|msg| eprintln!("{msg}")).await;
-        match adapter {
-            Some(adapter) => Arc::new(adapter),
-            None => {
-                // don't use eprintln! so test harness does not capture it
-                let _ = writeln!(
-                    std::io::stderr(),
-                    "Skipping rendering tests due to lack of wgpu::Adapter."
-                );
-                // Exit the process to skip redundant reports and make it clear that the
-                // tests aren't just passing
-                std::process::exit(0);
-            }
-        }
-    })
-    .await
-    .clone()
+pub(crate) async fn instance() -> &'static wgpu::Instance {
+    static CELL: tokio::sync::OnceCell<wgpu::Instance> = tokio::sync::OnceCell::const_new();
+    CELL.get_or_init(init::create_instance_for_test_or_exit)
+        .await
 }
 
 /// TODO: image probably isn't the best output format, just what I prototyped with
 pub(crate) async fn run_shader_test(test_wgsl: &str) -> image::Rgba32FImage {
-    let adapter = adapter().await;
+    let instance = instance().await;
+    let adapter = init::create_adapter_for_test(instance).await;
 
     // 32 is the minimum viewport width that will satisfy copy alignment
     let output_viewport = camera::Viewport::with_scale(1.0, [32, 32]);